+++ example-app/app.js
@@ -17,7 +17,7 @@ const server = HTTPS_CERTS ? https.createServer(HTTPS_CERTS, app) : http.createS
 
 // const expressOasGenerator = require('express-oas-generator')
 // expressOasGenerator.handleResponses(app, {})
-require('./common-express/preRoute')(app, global.CONFIG)
+require('@es-labs/node/express/preRoute')(app, express, global.CONFIG)
 
 // START SERVICES
 const { sleep } = require('esm')(module)('@es-labs/esm/sleep')
@@ -71,6 +71,6 @@ try {
 // https://github.com/mpashkovskiy/express-oas-generator/issues/24#issuecomment-764469904
 // expressOasGenerator.handleRequests()
 
-require('./common-express/postRoute')(app, express, global.CONFIG)
+require('@es-labs/node/express/postRoute')(app, express, global.CONFIG)
 
 module.exports = { server }